test: added testcase for slide snippets
