+++ slides/snippets/src/test/java/TestCreateBulletedText.java
@@ -1,0 +1,35 @@
+// Copyright 2022 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import com.google.api.services.slides.v1.model.BatchUpdatePresentationResponse;
+import org.junit.Test;
+
+import java.io.IOException;
+
+import static org.junit.Assert.assertEquals;
+
+// Unit testcase for createBulletedText snippet
+public class TestCreateBulletedText extends BaseTest{
+
+    @Test
+    public void testCreateBulletText() throws IOException {
+        String presentationId = createTestPresentation();
+        String pageId = createTestSlide(presentationId);
+        String boxId = createTestTextBox(presentationId, pageId);
+        BatchUpdatePresentationResponse response =
+                CreateBulletedText.createBulletedText(presentationId, boxId);
+        assertEquals(1, response.getReplies().size());
+        deleteFileOnCleanup(presentationId);
+    }
+}
