+++ slides/snippets/src/test/java/TestCreatePresentation.java
@@ -1,0 +1,30 @@
+// Copyright 2022 Google LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     https://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+import org.junit.Test;
+
+import java.io.IOException;
+
+import static org.junit.Assert.assertNotNull;
+
+// Unit testcase for createPresentation snippet
+public class TestCreatePresentation extends BaseTest{
+
+    @Test
+    public void testCreatePresentation() throws IOException {
+        String presentationId = CreatePresentation.createPresentation("Title");
+        assertNotNull(presentationId);
+        deleteFileOnCleanup(presentationId);
+    }
+}
